models/Emu3.5/requirements_new.txt
accelerate==1.11.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
einops==0.8.1
exceptiongroup==1.3.0
filelock==3.20.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
ninja==1.13.0
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12== 12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
packaging==25.0
pillow==11.3.0
pip==25.3
protobuf==6.33.0
psutil==7.1.3
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
setuptools==80.9.0
shellingham==1.5.4
sniffio==1.3.1
sympy==1.13.1
tiktoken==0.12.0
tokenizers==0.21.4
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.48.2
triton==3.1.0
typer-slim==0.20.0
typing_extensions==4.15.0
urllib3==2.5.0
wheel==0.45.1
